Extracted we_fl_unix.h from edit.h
