+++ we_fl_fkt.h
@@ -3,5 +3,6 @@
 
 #include "we_edit.h"
 #include "we_e_aus.h"
+#include "we_fl_unix.h"
 
 #endif
