+++ we_main.h
@@ -4,6 +4,7 @@
 #include "globals.h"
 #include "we_edit.h"
 #include "we_e_aus.h"
+#include "we_fl_unix.h"
 
 void e_ini_desk(ECNT * cn);
 void FARBE_Init(FARBE * fb);
